Include instead of inheritance
