+++ lib/goodmin.rb
@@ -1,6 +1,7 @@
 require "importmap-rails"
 require "stimulus-rails"
 require "csv_builder"
+require "goodmin/application_controller"
 require "goodmin/authentication"
 require "goodmin/authorization"
 require "goodmin/engine"
@@ -25,6 +26,7 @@ require "goodmin/helpers/navigation"
 require "goodmin/helpers/translations"
 require "goodmin/helpers/filters"
 require "goodmin/helpers/tables"
+require "goodmin/resources/resource_controller"
 require "goodmin/resources/resource"
 require "goodmin/resources/resource_service"
 require "goodmin/resources/form_component"
